database/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
grpcio==1.70.0
grpcio-tools==1.70.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
protobuf==5.29.0
Werkzeug==3.0.1
Flask-CORS==4.0.0
watchdog==6.0.0